Add Exception and Code Refactor

+++ springboot/curry/src/main/java/com/ottugi/curry/service/user/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package com.ottugi.curry.service.user;
 
 import com.ottugi.curry.domain.user.*;
+import com.ottugi.curry.service.CommonService;
 import com.ottugi.curry.web.dto.user.UserResponseDto;
 import com.ottugi.curry.web.dto.user.UserUpdateRequestDto;
 import lombok.RequiredArgsConstructor;
@@ -8,25 +9,25 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-@Transactional
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
+    private final CommonService commonService;
 
     // 회원 정보 조회
     @Override
+    @Transactional(readOnly = true)
     public UserResponseDto getProfile(Long id) {
-
-        User user = findUser(id);
+        User user = commonService.findByUserId(id);
         return new UserResponseDto(user);
     }
 
     // 회원 정보 수정
     @Override
+    @Transactional
     public UserResponseDto updateProfile(UserUpdateRequestDto userUpdateRequestDto) {
-
-        User user = findUser(userUpdateRequestDto.getId());
+        User user = commonService.findByUserId(userUpdateRequestDto.getId());
         user.updateProfile(userUpdateRequestDto.getNickName());
         return new UserResponseDto(user);
     }
@@ -34,15 +35,11 @@ public class UserServiceImpl implements UserService {
     // 회원 탈퇴
     @Override
     public Boolean setWithdraw(Long id) {
-
-        User user = findUser(id);
-        userRepository.delete(user);
-        return true;
-    }
-
-    // 회원 조회
-    public User findUser(Long id) {
-
-        return userRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 회원이 없습니다."));
+        User user = commonService.findByUserId(id);
+        if (user != null) {
+            userRepository.delete(user);
+            return true;
+        }
+        return false;
     }
 }
